+++ app/src/main/java/com/example/rememberall/ui/main/MainActivity.kt
@@ -6,16 +6,11 @@ import androidx.activity.compose.setContent
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
-import androidx.compose.material3.TopAppBarDefaults.topAppBarColors
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
-import com.example.rememberall.ui.components.appbar.AppBarBack
 import com.example.rememberall.ui.navigation.AppNavigationHost
 import com.example.rememberall.ui.theme.RememberAllTheme
 import dagger.hilt.android.AndroidEntryPoint
