+++ app/src/main/java/com/example/rememberall/ui/navigation/destinations/MainDestination.kt
@@ -1,0 +1,28 @@
+package com.example.rememberall.ui.navigation.destinations
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.NavHostController
+import com.example.rememberall.ui.base.components.BaseDestination
+import com.example.rememberall.ui.main.MainContract
+import com.example.rememberall.ui.main.MainScreen
+import com.example.rememberall.ui.main.MainViewModel
+
+private val MAIN_SCREEN_EFFECT_KEY = "main_screen_effect_key"
+
+@Composable
+fun MainDestination(navController: NavHostController) =
+    BaseDestination<MainContract.State, MainContract.Effect, MainContract.Intent, MainViewModel> { state, effect, effectListener, intentListener ->
+
+        LaunchedEffect(MAIN_SCREEN_EFFECT_KEY) {
+            effect.collect {
+                when(it)
+                {
+                    else -> {}
+                }
+            }
+        }
+
+        MainScreen(state, effectListener, intentListener)
+    }
