+++ app/src/main/java/com/example/rememberall/ui/navigation/destinations/NotesDestination.kt
@@ -1,15 +1,18 @@
-package com.example.rememberall.ui.notes
+package com.example.rememberall.ui.navigation.destinations
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.navigation.NavHostController
 import com.example.rememberall.ui.base.components.BaseDestination
 import com.example.rememberall.ui.navigation.NoteDetailRoute
+import com.example.rememberall.ui.notes.NotesContract
+import com.example.rememberall.ui.notes.NotesScreen
+import com.example.rememberall.ui.notes.NotesViewModel
 
 private val NOTE_SCREEN_EFFECT_KEY = "note_screen_effect_key"
 
 @Composable
-fun NoteDestination(navController: NavHostController) {
+fun NoteDestination(navController: NavHostController)  =
     BaseDestination<NotesContract.State, NotesContract.Effect, NotesContract.Intent, NotesViewModel> { state, effect, effectListener, intentListener ->
 
         LaunchedEffect(NOTE_SCREEN_EFFECT_KEY) {
@@ -24,4 +27,3 @@ fun NoteDestination(navController: NavHostController) {
 
         NotesScreen(state, effectListener, intentListener)
     }
-}
