+++ app/src/main/java/com/example/rememberall/ui/theme/Theme.kt
@@ -15,7 +15,6 @@ import androidx.core.view.WindowCompat
 @Composable
 fun RememberAllTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
-    // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true, content: @Composable () -> Unit
 )
 {
